Added the Graph Feature to look at historical data for perf
Can view the test and baseline scores, as well as the regression percentages for any specified date range.

How to: In Tabular View, clicking on any cell will lead to the Perf Compare page. On this page there is a Graph button that will lead to the graph page.

Remove data points feature is not fully working yet. You should be able to click any data point to remove it.

Signed-off-by: Manasha Vetrivelu <Manasha.Vetrivelu@ibm.com>